+++ packages/utils/examples/utils/pages/labels/labels.page.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+	templateUrl: './labels.page.html',
+})
+export class UtilsLabelsDemoPageComponent {
+
+}